+++ src/main/java/dev/krk/emulator/chipsekiz/opcodes/Op1NNN.java
@@ -5,7 +5,7 @@ package dev.krk.emulator.chipsekiz.opcodes;
  * Jumps to address NNN.
  */
 public class Op1NNN extends OpPNNN {
-    public Op1NNN(int address) {
+    public Op1NNN(short address) {
         super(1, address);
     }
 }
