+++ src/main/java/dev/krk/emulator/chipsekiz/opcodes/Op2NNN.java
@@ -4,7 +4,7 @@ package dev.krk.emulator.chipsekiz.opcodes;
  * Calls subroutine at NNN.
  */
 public class Op2NNN extends OpPNNN {
-    public Op2NNN(int address) {
+    public Op2NNN(short address) {
         super(2, address);
     }
 }
